requirements.txt
mysql-connector-python
numpy
matplotlib
tkinter
random